+++ bot.py
@@ -201,87 +201,6 @@ def getRandomJoke():
     return getJoke()
 
 
-def sureWantTo(update, context):
-    user = update.message.from_user
-    message = update.message.text
-    print(message)
-    if message == "YES":
-        update.message.reply_text("Goodbye!", reply_markup=create_markup([["/start"]]))
-        users.delete_one({"userId": user.id})
-        return ConversationHandler.END
-
-    if message == "NO":
-        update.message.reply_text(
-            "Oh thanks god, I thought I lost you",
-            reply_markup=get_reply_keyboard_markup(),
-        )
-        return LOOP
-
-
-def loopAnswer(update, context):
-    user = update.message.from_user
-    info = mongo.getUserInfo(user.id)
-    print(info)
-    message = update.message.text
-    reply = ""
-    if message == "💰 Balance":
-        refbal = "{:,.2f}".format(info["refCount"] * REFERRAL_REWARD)
-        if refbal == "":
-            refbal = "0"
-        reply = BALANCE_TEXT.replace("IARTBALANCE", AIRDROP_AMOUNT).replace(
-            "REFERRALBALANCE", refbal
-        )
-
-    if message == "ℹ️ Airdrop Info":
-        reply = PROCEED_MESSAGE
-
-    if message == "💸 Withdrawal":
-        reply = WITHDRAWAL_TEXT
-
-    if message == "🔗 Ref Link":
-        reply = f"""
-Here is *your referral link*
-[https://t.me/{context.bot.username}?start={user.id}](https://t.me/{context.bot.username}?start={user.id})
-"""
-
-    if message == "Quit Airdrop":
-        update.message.reply_text(
-            "Are you sure want to quit the Airdrop? All your data will be deleted",
-            reply_markup=create_markup([["YES"], ["NO"]]),
-        )
-        return SUREWANTTO
-
-    if message == "💾 My Data":
-        name = str(info["name"])
-        refbal = "{:,.2f}".format(info["refCount"] * REFERRAL_REWARD)
-        balance = BALANCE_TEXT.replace("IARTBALANCE", AIRDROP_AMOUNT).replace(
-            "REFERRALBALANCE", refbal
-        )
-        refferals = str(info["refCount"])
-        bep20Address = str(info["bep20"])
-        twitterUsername = str(info["twitter_username"])
-        reply = f"""
-Name: {name}
-Referrals: {refferals}
-{AIRDROP_NETWORK} address: {bep20Address}
-Twitter Username: {twitterUsername}
-"""
-    if reply == "":
-        joke = getRandomJoke()
-        joke = joke.split("  -")
-        reply = f"""
-I'm not sure what you meant, but here is a joke for you!
-> {joke[0]}
-- {joke[1]}
-"""
-    update.message.reply_text(
-        reply,
-        reply_markup=get_reply_keyboard_markup(),
-        parse_mode=telegram.ParseMode.MARKDOWN,
-    )
-    return LOOP
-
-
 # %% Start bot
 
 cancelHandler = MessageHandler(Filters.regex("^Cancel$"), cancel)
